Extract common logic for detectStart and detectStop into a helper class.
