+++ src/BodyPose/index.js
@@ -8,8 +8,9 @@ BodyPose
 Ported from pose-detection at Tensorflow.js
 */
 
-import * as tf from "@tensorflow/tfjs";
 import * as poseDetection from "@tensorflow-models/pose-detection";
+import * as tf from "@tensorflow/tfjs";
+import ImageDetector from "../ImageDetector";
 import callCallback from "../utils/callcallback";
 import handleArguments from "../utils/handleArguments";
 import { mediaReady } from "../utils/imageUtilities";
@@ -55,13 +56,6 @@ class BodyPose {
     this.model = null;
     this.config = options;
     this.runtimeConfig = {};
-    this.detectMedia = null;
-    this.detectCallback = null;
-
-    // flags for detectStart() and detectStop()
-    this.detecting = false; // true when detection loop is running
-    this.signalStop = false; // Signal to stop the loop
-    this.prevCall = ""; // Track previous call to detectStart() or detectStop()
 
     this.ready = callCallback(this.loadModel(), callback);
   }
@@ -129,105 +123,23 @@ class BodyPose {
     this.model = await poseDetection.createDetector(pipeline, modelConfig);
 
     // for compatibility with p5's preload()
-    if (this.p5PreLoadExists) window._decrementPreload();
+    if (this.p5PreLoadExists()) window._decrementPreload();
 
     return this;
   }
 
   /**
-   * A callback function that handles the pose detection results.
-   * @callback gotPoses
-   * @param {Array} results - An array of objects containing poses.
-   */
-
-  /**
    * Asynchronously outputs a single pose prediction result when called.
-   * @param {*} media - An HMTL or p5.js image, video, or canvas element to run the prediction on.
-   * @param {gotPoses} callback - A callback function to handle the predictions.
+   * @param {*} media - An HTML or p5.js image, video, or canvas element to run the prediction on.
    * @returns {Promise<Array>} an array of poses.
    */
-  async detect(...inputs) {
-    //Parse out the input parameters
-    const argumentObject = handleArguments(...inputs);
-    argumentObject.require(
-      "image",
-      "An html or p5.js image, video, or canvas element argument is required for detect()."
-    );
-    const { image, callback } = argumentObject;
-
-    await mediaReady(image, false);
+  async detect(media) {
+    await mediaReady(media, false);
     const predictions = await this.model.estimatePoses(
-      image,
+      media,
       this.runtimeConfig
     );
-    let result = predictions;
-    result = this.addKeypoints(result);
-    if (typeof callback === "function") callback(result);
-    return result;
-  }
-
-  /**
-   * Repeatedly outputs pose predictions through a callback function.
-   * Calls the internal detectLoop() function.
-   * @param {*} media - An HMTL or p5.js image, video, or canvas element to run the prediction on.
-   * @param {gotPoses} callback - A callback function to handle the predictions.
-   * @returns {Promise<Array>} an array of predictions.
-   */
-  detectStart(...inputs) {
-    // Parse out the input parameters
-    const argumentObject = handleArguments(...inputs);
-    argumentObject.require(
-      "image",
-      "An html or p5.js image, video, or canvas element argument is required for detectStart()."
-    );
-    argumentObject.require(
-      "callback",
-      "A callback function argument is required for detectStart()."
-    );
-    this.detectMedia = argumentObject.image;
-    this.detectCallback = argumentObject.callback;
-
-    this.signalStop = false;
-    if (!this.detecting) {
-      this.detecting = true;
-      this.detectLoop();
-    }
-    if (this.prevCall === "start") {
-      console.warn(
-        "detectStart() was called more than once without calling detectStop(). The lastest detectStart() call will be used and the previous calls will be ignored."
-      );
-    }
-    this.prevCall = "start";
-  }
-
-  /**
-   * Internal function that calls estimatePoses in a loop
-   * Can be started by detectStart() and terminated by detectStop()
-   * @private
-   */
-  async detectLoop() {
-    await mediaReady(this.detectMedia, false);
-    while (!this.signalStop) {
-      const predictions = await this.model.estimatePoses(
-        this.detectMedia,
-        this.runtimeConfig
-      );
-      let result = predictions;
-      result = this.addKeypoints(result);
-      this.detectCallback(result);
-      // wait for the frame to update
-      await tf.nextFrame();
-    }
-    this.detecting = false;
-    this.signalStop = false;
-  }
-
-  /**
-   * Stops the detection loop before next detection loop runs.
-   */
-  detectStop() {
-    if (this.detecting) this.signalStop = true;
-    this.prevCall = "stop";
+    return this.addKeypoints(predictions);
   }
 
   /**
@@ -268,12 +180,12 @@ class BodyPose {
 
 /**
  * Factory function that returns a BodyPose instance.
- * @returns {BodyPose} A BodyPose instance.
+ * @returns {ImageDetector} A BodyPose instance.
  */
 const bodyPose = (...inputs) => {
   const { string, options = {}, callback } = handleArguments(...inputs);
   const instance = new BodyPose(string, options, callback);
-  return instance;
+  return new ImageDetector(instance);
 };
 
 export default bodyPose;
