+++ src/FaceMesh/index.js
@@ -10,16 +10,19 @@
 
 import * as tf from "@tensorflow/tfjs";
 import * as faceLandmarksDetection from "@tensorflow-models/face-landmarks-detection";
+import ImageDetector from "../ImageDetector";
 import callCallback from "../utils/callcallback";
 import handleArguments from "../utils/handleArguments";
-import { mediaReady } from "../utils/imageUtilities";
 
+/**
+ * @implements {SpecificDetectorImplementation}
+ */
 class FaceMesh {
   /**
    * An options object to configure FaceMesh settings
    * @typedef {Object} configOptions
-   * @property {number} maxFacess - The maximum number of faces to detect. Defaults to 2.
-   * @property {boolean} refineLandmarks - Refine the ladmarks. Defaults to false.
+   * @property {number} maxFaces - The maximum number of faces to detect. Defaults to 2.
+   * @property {boolean} refineLandmarks - Refine the landmarks. Defaults to false.
    * @property {boolean} flipHorizontal - Flip the result horizontally. Defaults to false.
    * @property {string} runtime - The runtime to use. "mediapipe"(default) or "tfjs".
    *
@@ -41,13 +44,6 @@ class FaceMesh {
     this.model = null;
     this.config = options;
     this.runtimeConfig = {};
-    this.detectMedia = null;
-    this.detectCallback = null;
-
-    // flags for detectStart() and detectStop()
-    this.detecting = false; // true when detection loop is running
-    this.signalStop = false; // true when detectStop() is called and detecting is true
-    this.prevCall = ""; // "start" or "stop", used for giving warning messages with detectStart() is called twice in a row
 
     this.ready = callCallback(this.loadModel(), callback);
   }
@@ -78,99 +74,22 @@ class FaceMesh {
     );
 
     // for compatibility with p5's preload()
-    if (this.p5PreLoadExists) window._decrementPreload();
+    if (this.p5PreLoadExists()) window._decrementPreload();
 
     return this;
   }
 
   /**
    * Asynchronously output a single face prediction result when called
-   * @param {*} [media] - An HMTL or p5.js image, video, or canvas element to run the prediction on.
-   * @param {function} [callback] - A callback function to handle the predictions.
+   * @param {*} [media] - An HTML or p5.js image, video, or canvas element to run the prediction on.
    * @returns {Promise<Array>} an array of predictions.
    */
-  async detect(...inputs) {
-    // Parse out the input parameters
-    const argumentObject = handleArguments(...inputs);
-    argumentObject.require(
-      "image",
-      "An html or p5.js image, video, or canvas element argument is required for detect()."
-    );
-    const { image, callback } = argumentObject;
-
-    await mediaReady(image, false);
+  async detect(media) {
     const predictions = await this.model.estimateFaces(
-      image,
+      media,
       this.runtimeConfig
     );
-    let result = predictions;
-    result = this.addKeypoints(result);
-    if (typeof callback === "function") callback(result);
-    return result;
-  }
-
-  /**
-   * Repeatedly output face predictions through a callback function
-   * @param {*} [media] - An HMTL or p5.js image, video, or canvas element to run the prediction on.
-   * @param {function} [callback] - A callback function to handle the predictions.
-   * @returns {Promise<Array>} an array of predictions.
-   */
-  detectStart(...inputs) {
-    // Parse out the input parameters
-    const argumentObject = handleArguments(...inputs);
-    argumentObject.require(
-      "image",
-      "An html or p5.js image, video, or canvas element argument is required for detectStart()."
-    );
-    argumentObject.require(
-      "callback",
-      "A callback function argument is required for detectStart()."
-    );
-    this.detectMedia = argumentObject.image;
-    this.detectCallback = argumentObject.callback;
-
-    this.signalStop = false;
-    if (!this.detecting) {
-      this.detecting = true;
-      this.detectLoop();
-    }
-    if (this.prevCall === "start") {
-      console.warn(
-        "detectStart() was called more than once without calling detectStop(). The lastest detectStart() call will be used and the previous calls will be  ignored."
-      );
-    }
-    this.prevCall = "start";
-  }
-
-  /**
-   * Stop the detection loop before next detection loop runs.
-   */
-  detectStop() {
-    if (this.detecting) this.signalStop = true;
-    this.prevCall = "stop";
-  }
-
-  /**
-   * Internal function to call estimateFaces in a loop
-   * Can be started by detectStart() and terminated by detectStop()
-   *
-   * @private
-   */
-  async detectLoop() {
-    await mediaReady(this.detectMedia, false);
-    while (!this.signalStop) {
-      const predictions = await this.model.estimateFaces(
-        this.detectMedia,
-        this.runtimeConfig
-      );
-      let result = predictions;
-      result = this.addKeypoints(result);
-      this.detectCallback(result);
-      // wait for the frame to update
-      await tf.nextFrame();
-    }
-    this.detecting = false;
-    this.signalStop = false;
+    return this.addKeypoints(predictions);
   }
 
   /**
@@ -265,12 +184,12 @@ class FaceMesh {
 
 /**
  * Factory function that returns a FaceMesh instance
- * @returns {Object} A new faceMesh instance
+ * @returns {ImageDetector} A new faceMesh instance
  */
 const faceMesh = (...inputs) => {
   const { options = {}, callback } = handleArguments(...inputs);
   const instance = new FaceMesh(options, callback);
-  return instance;
+  return new ImageDetector(instance);
 };
 
 export default faceMesh;
